requirements.txt
class-resolver
einops
fire
gin-config
numpy
pandas
prettytable
pyro-ppl
scikit-learn
torch-geometric
torch-scatter
torch-sparse
torchmetrics
torchvision
tqdm
wandb